refactored server layout and return codes
