+++ server/package.json
@@ -5,7 +5,7 @@
 	"main": "index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node index.js"
+		"start": "node src/index.js"
 	},
 	"author": "",
 	"license": "ISC",
